Hand detection
Resolved the inverted histogram mask problem.
The script can now based on a preliminar skin color HSV detection and offsets detect the hand even if the brightness inc/dec.

To-Do
Filter the histogrammed result to capture the entire hand without holes.
Clean the noise in the rest of the image e.g.: the face is considered the largest contour